Hang on to git_repository until all objects are closed.
Based on [1] and [2] 'git_close_object object >> touchForeignPtr
repository' seems to be a safe way to keep the repo alive long
enough. Even if GC triggers the finalizer for git_repository_free at
the same time as triggering the finalizer for the object the close of
object should not be getting any attributes and both just result in
decreasing the object reference counts.

[1] from libgit2/**/repository.h:

Note that after a repository is free'd, all the objects it has spawned
will still exist until they are manually closed by the user with
`git_object_close`, but accessing any of the attributes of an object
without a backing repository will result in undefined behavior

[2] from docs for ForeignPtr:

Note that this function should not be used to express dependencies
between finalizers on ForeignPtrs. For example, if the finalizer for a
ForeignPtr F1 calls touchForeignPtr on a second ForeignPtr F2, then
the only guarantee is that the finalizer for F2 is never started
before the finalizer for F1. They might be started together if for
example both F1 and F2 are otherwise unreachable, and in that case the
scheduler might end up running the finalizer for F2 first.